Refactored blogs.ts page
