+++ blogs/blogs.css
@@ -1,48 +1,3 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.user-details {
-  width: 250px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  border-color: #3333;
-  border-style: solid;
-  border-width: thin;
-  border-left: none;
-  border-right: none;
-  padding: 0.5em 0;
-}
-.user-details > p {
-  text-transform: capitalize;
-  padding-bottom: 0.5em;
-  letter-spacing: 0.1em;
-  font-weight: 500;
-}
-.user-details div {
-  display: flex;
-  align-items: center;
-}
-.user-details img {
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  margin-right: 1em;
-}
-
-.logout {
-  padding: 0.5em 1em;
-  margin: 0 0.3em;
-  color: #fff;
-  cursor: pointer;
-  background: #3c6e71ff;
-  border-radius: 5px;
-}
-
 .blogs {
   width: 90%;
   display: flex;
