+++ create/index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Ultimate Blog App | Welcome</title>
     <link rel="stylesheet" href="../global.css" />
-    <link rel="stylesheet" href="blogs.css" />
+    <link rel="stylesheet" href="create.css" />
   </head>
   <body>
     <main>
@@ -20,14 +20,12 @@
             <p>Hello, <span class="username"></span></p>
             <a href="../" class="logout">Logout</a>
           </div>
-          <a class="link-page" href="../create/">Create new article </a>
+          <a class="link-page" href="../blogs/">View all articles </a>
         </div>
-
-        <!-- blog list -->
-        <div class="blogs"></div>
+        <h1>New article</h1>
       </div>
     </main>
 
-    <script type="module" src="../public/blogs.js"></script>
+    <script src="../public/create.js" type="module"></script>
   </body>
 </html>
