+++ global.css
@@ -35,3 +35,63 @@ h1 {
   letter-spacing: 1.5px;
   text-align: center;
 }
+
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.user-details {
+  width: 250px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  border-color: #3333;
+  border-style: solid;
+  border-width: thin;
+  border-left: none;
+  border-right: none;
+  padding: 0.5em 0;
+}
+.user-details > p {
+  text-transform: capitalize;
+  padding-bottom: 0.5em;
+  letter-spacing: 0.1em;
+  font-weight: 500;
+}
+.user-details div {
+  display: flex;
+  align-items: center;
+}
+.user-details img {
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  margin-right: 1em;
+}
+
+.logout {
+  padding: 0.5em 1em;
+  margin: 0 0.3em;
+  color: #fff;
+  cursor: pointer;
+  background: #3c6e71ff;
+  border-radius: 5px;
+}
+
+.link-page {
+  line-height: 2em;
+  font-weight: 300;
+  position: relative;
+}
+.link-page::after {
+  content: "";
+  position: absolute;
+  border-bottom: #3333 solid 1px;
+  width: 100%;
+  height: 100%;
+  bottom: 0.3em;
+  left: 0;
+}
