+++ public/create.js
@@ -1,0 +1,3 @@
+import uuid from "./utils/uuid.js";
+console.log("uuid", uuid());
+// console.log(JSON.parse(data));
