+++ public/models/post.js
@@ -1,0 +1,1 @@
+export {};
