+++ public/models/user.js
@@ -1,0 +1,1 @@
+export {};
