+++ public/utils/logout.js
@@ -1,0 +1,5 @@
+const logout = () => {
+    localStorage.removeItem("user");
+    window.location.href = "../";
+};
+export default logout;
