+++ public/utils/models/userDetails.js
@@ -1,0 +1,1 @@
+export {};
