+++ src/create.ts
@@ -1,0 +1,6 @@
+import uuid from "./utils/uuid.js";
+import posts from "./data/posts.js";
+
+console.log("uuid", uuid());
+
+// console.log(JSON.parse(data));
