+++ src/models/loginDetails.ts
@@ -1,0 +1,4 @@
+export default interface LOGINDETAILS {
+  username: string;
+  password: string;
+}
