+++ src/models/post.ts
@@ -1,0 +1,6 @@
+export default interface POST {
+  userId: string;
+  id: string;
+  title: string;
+  body: string;
+}
