+++ src/models/user.ts
@@ -1,0 +1,7 @@
+export default interface USER {
+  id: string;
+  firstName: string;
+  lastName: string;
+  username: string;
+  password: string;
+}
