+++ src/models/userDetails.ts
@@ -1,0 +1,5 @@
+export default interface USERDETAILS {
+  firstname: string;
+  lastname: string;
+  username: string;
+}
