+++ src/utils/logout.ts
@@ -1,0 +1,7 @@
+const logout = () => {
+  localStorage.removeItem("user");
+
+  window.location.href = "../";
+};
+
+export default logout;
